first_project/requirements.txt
itsdangerous~=2.1.2
Django~=4.1.3
djangorestframework~=3.14.0
django-filters